+++ CHANGELOG.md
@@ -40,3 +40,7 @@
 
 - [x] node#copy :: 파일 복사 구현.
 - [x] Promise function 이름에 _ 를 지우자.
+- [x] debug#createSpy()에 function을 집어 넣을 수 있고,
+  function을 대신 실행하고 결과 값을 저장해 둔다.
+- [x] debug#createSpy(): 결과 값 저장할때 unshift 순으로 저장한다.
+  고로 가장 최근 저장된 결과 값은 returns[0]일 것이다.